requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
python-multipart
pyjwt
pwdlib
httpx
pytest
pydantic[email]
pwdlib[argon2]